Move type utilities into a separate file
Also adds detailed comments.
