+++ stablehlo/reference/CMakeLists.txt
@@ -12,6 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+add_mlir_library(StablehloReferenceTypes
+  PARTIAL_SOURCES_INTENDED
+  Types.cpp
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+)
+
 add_mlir_library(StablehloReferenceElement
   PARTIAL_SOURCES_INTENDED
   Element.cpp
@@ -19,6 +27,7 @@ add_mlir_library(StablehloReferenceElement
   LINK_LIBS PUBLIC
   MLIRIR
   MLIRSupport
+  StablehloReferenceTypes
 )
 
 add_mlir_library(StablehloReferenceIndex
@@ -38,6 +47,7 @@ add_mlir_library(StablehloReferenceTensor
   MLIRIR
   StablehloReferenceElement
   StablehloReferenceIndex
+  StablehloReferenceTypes
 )
 
 add_mlir_library(StablehloReferenceOps
